Frontend - Memories Projects - Part 4
Card:
Title was not showing, corrected that.

Also, when editing in the inputs card the title "Creating a Memory" should change when we're editing, so using the currentId as a verification step, if we have the currentId will display "Edit Memory" instead.

Form:
<Typography variant='h6'>Creating a Memory</Typography>
to:
<Typography variant='h6'>{currentId ? "Editing" : "Creating"} a Memory</Typography>

Also, when inputs or changes are done, all inputs should be clear and the form should go back to the "default" display with the title, "Creating a Memory", so we implemented the "clear" function, and we're calling on both scenarios, if the user is editing a memory or creating one.